+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- new glob exclude pattern `!`
+
 ## 1.3.1 - 2020-05-15
 
 - fixed removing files on Windows